requirements.txt
numpy
torch
matplotlib
alpha_vantage
python-dotenv